+++ tests/_self/duplicate-test-name.js
@@ -1,0 +1,21 @@
+const test = require('../../aqa')
+
+test('T1', t => {
+    t.true(true);
+})
+
+test('T2', t => {
+    t.true(true);
+})
+
+test('T1', t => {
+    t.true(true);
+})
+
+test('T2', t => {
+    t.true(true);
+})
+
+test('T3', t => {
+    t.true(true);
+})