+++ tests/_self/log.js
@@ -1,0 +1,9 @@
+const test = require('../../aqa')
+
+test('Test 1', t => {
+    t.log('Hello');
+})
+
+test('Test 2', t => {
+    t.log('World');
+})