+++ tests/_self/throw.js
@@ -1,0 +1,9 @@
+const test = require('../../aqa')
+
+function sut() {
+    throw new Error('Bye')
+}
+
+test('Throw test', _t => {
+    sut();
+})